fix: move shared types to separate file
