+++ src/shared/utils/@types/diagnose.ts
@@ -1,0 +1,2 @@
+export type Kind = "raw" | "wrapped" | "signed";
+export type Mode = "strict" | "non-strict";
