+++ src/shared/utils/diagnose.ts
@@ -9,10 +9,9 @@ import {
 import { ArrayProof, Signature, SignatureStrict } from "../../2.0/types";
 import { clone, cloneDeepWith } from "lodash";
 import { buildAjv, getSchema } from "../ajv";
+import { Kind, Mode } from "./@types/diagnose";
 
 type Version = "2.0" | "3.0";
-export type Kind = "raw" | "wrapped" | "signed";
-export type Mode = "strict" | "non-strict";
 
 interface DiagnoseError {
   message: string;
